First version with interpolation always on
