+++ Client/Systems/SettingsSys/SettingsStructures.cs
@@ -22,7 +22,6 @@ namespace LunaClient.Systems.SettingsSys
         public List<ServerEntry> Servers { get; set; } = new List<ServerEntry>();
         public int InitialConnectionSyncTimeRequests { get; set; } = 10;
         public bool RevertEnabled { get; set; }
-        public bool InterpolationEnabled { get; set; } = false;
         public int MaxGroupsPerPlayer { get; set; } = 1;
         public bool OverrideIntegrator { get; set; }
