+++ Client/Systems/VesselPositionSys/VesselPositionMessageHandler.cs
@@ -1,10 +1,10 @@
 ﻿using LunaClient.Base;
 using LunaClient.Base.Interface;
-using LunaClient.Systems.SettingsSys;
 using LunaClient.VesselStore;
 using LunaClient.VesselUtilities;
 using LunaCommon.Message.Data.Vessel;
 using LunaCommon.Message.Interface;
+using System;
 using System.Collections.Concurrent;
 
 namespace LunaClient.Systems.VesselPositionSys
@@ -18,44 +18,45 @@ namespace LunaClient.Systems.VesselPositionSys
             if (!(msg.Data is VesselPositionMsgData msgData)) return;
             
             var vesselId = msgData.VesselId;
-
             if (!VesselCommon.DoVesselChecks(vesselId))
                 return;
 
             //Vessel might exist in the store but not in game (if the vessel is in safety bubble for example)
             VesselsProtoStore.UpdateVesselProtoPosition(msgData);
 
-            if (!VesselPositionSystem.CurrentVesselUpdate.TryGetValue(vesselId, out var existingPositionUpdate))
+            var update = CreatePosUpdateFromMessage(msgData);
+
+            if (!VesselPositionSystem.CurrentVesselUpdate.ContainsKey(vesselId))
             {
-                var current = MessageToPositionTransfer.CreateFromMessage(msg);
-                var target = MessageToPositionTransfer.CreateFromMessage(msg);
-
-                VesselPositionSystem.CurrentVesselUpdate.TryAdd(vesselId, current);
-                VesselPositionSystem.TargetVesselUpdate.TryAdd(vesselId, target);
+                VesselPositionSystem.CurrentVesselUpdate.TryAdd(vesselId, update);
                 VesselPositionSystem.TargetVesselUpdateQueue.TryAdd(vesselId, new FixedSizedConcurrentQueue<VesselPositionUpdate>(VesselPositionSystem.MaxQueuedUpdates));
-
-                current.SetTarget(target);
             }
             else
             {
-                if (SettingsSystem.CurrentSettings.InterpolationEnabled)
+                if (VesselPositionSystem.TargetVesselUpdateQueue.TryGetValue(vesselId, out var queue))
                 {
-                    if (VesselPositionSystem.TargetVesselUpdateQueue.TryGetValue(vesselId, out var queue))
-                    {
-                        queue.Enqueue(MessageToPositionTransfer.CreateFromMessage(msg));
-                    }
-                }
-                else
-                {
-                    if (VesselPositionSystem.TargetVesselUpdate.TryGetValue(vesselId, out var existingTargetPositionUpdate))
-                    {                    
-                        //Overwrite the TARGET data with the data we've received in the message
-                        MessageToPositionTransfer.UpdateFromMessage(msg, existingTargetPositionUpdate);
-                        //Here we just set the interpolation as not started
-                        existingPositionUpdate.SetTarget(existingTargetPositionUpdate);
-                    }
+                    queue.Enqueue(update);
                 }
             }
+        }
+
+        private static VesselPositionUpdate CreatePosUpdateFromMessage(VesselPositionMsgData msgData)
+        {
+            var update = new VesselPositionUpdate
+            {
+                VesselId = msgData.VesselId,
+                BodyIndex = msgData.BodyIndex,
+                HeightFromTerrain = msgData.HeightFromTerrain,
+                GameTimeStamp = msgData.GameTime
+            };
+
+            Array.Copy(msgData.SrfRelRotation, update.SrfRelRotation, 4);
+            Array.Copy(msgData.Velocity, update.Velocity, 3);
+            Array.Copy(msgData.LatLonAlt, update.LatLonAlt, 3);
+            Array.Copy(msgData.NormalVector, update.NormalVector, 3);
+            Array.Copy(msgData.Orbit, update.Orbit, 8);
+
+            return update;
         }
     }
 }
