+++ Common/Message/Data/Vessel/VesselPositionMsgData.cs
@@ -16,10 +16,7 @@ namespace LunaCommon.Message.Data.Vessel
         public int BodyIndex;
         public double[] LatLonAlt = new double[3];
         public double[] NormalVector = new double[3];
-        public double[] TransformPosition = new double[3];
         public double[] Velocity = new double[3];
-        public double[] OrbitPos = new double[3];
-        public double[] OrbitVel = new double[3];
         public double[] Orbit = new double[8];
         public float[] SrfRelRotation = new float[4];
         public float HeightFromTerrain;
@@ -42,16 +39,7 @@ namespace LunaCommon.Message.Data.Vessel
                 lidgrenMsg.Write(NormalVector[i]);
 
             for (var i = 0; i < 3; i++)
-                lidgrenMsg.Write(TransformPosition[i]);
-
-            for (var i = 0; i < 3; i++)
                 lidgrenMsg.Write(Velocity[i]);
-
-            for (var i = 0; i < 3; i++)
-                lidgrenMsg.Write(OrbitPos[i]);
-
-            for (var i = 0; i < 3; i++)
-                lidgrenMsg.Write(OrbitVel[i]);
 
             for (var i = 0; i < 8; i++)
                 lidgrenMsg.Write(Orbit[i]);
@@ -76,18 +64,9 @@ namespace LunaCommon.Message.Data.Vessel
             
             for (var i = 0; i < 3; i++)
                 NormalVector[i] = lidgrenMsg.ReadDouble();
-
-            for (var i = 0; i < 3; i++)
-                TransformPosition[i] = lidgrenMsg.ReadDouble();
             
             for (var i = 0; i < 3; i++)
                 Velocity[i] = lidgrenMsg.ReadDouble();
-
-            for (var i = 0; i < 3; i++)
-                OrbitPos[i] = lidgrenMsg.ReadDouble();
-
-            for (var i = 0; i < 3; i++)
-                OrbitVel[i] = lidgrenMsg.ReadDouble();
 
             for (var i = 0; i < 8; i++)
                 Orbit[i] = lidgrenMsg.ReadDouble();
@@ -102,7 +81,7 @@ namespace LunaCommon.Message.Data.Vessel
         
         internal override int InternalGetMessageSize()
         {
-            return base.InternalGetMessageSize() + GuidUtil.GetByteSize() + sizeof(int) + sizeof(double) * 3 * 6 + 
+            return base.InternalGetMessageSize() + GuidUtil.GetByteSize() + sizeof(int) + sizeof(double) * 3 * 3 + 
                 sizeof(float) * 4 * 1 + sizeof(float) + sizeof(long) + sizeof(double);
         }
     }
