+++ LunaMultiPlayer.sln
@@ -14,6 +14,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 		appveyor.yml = appveyor.yml
 		Scripts\CopyToKSPDirectory.bat = Scripts\CopyToKSPDirectory.bat
+		README.md = README.md
 		Scripts\SetDirectories.bat = Scripts\SetDirectories.bat
 	EndProjectSection
 EndProject
