+++ LunaUpdater/Properties/AssemblyInfo.cs
@@ -13,6 +13,6 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("35a57b77-236e-492c-b003-1f8fc2f544f3")]
 
-[assembly: AssemblyVersion("0.6.37")]
-[assembly: AssemblyFileVersion("0.6.37")]
-[assembly: AssemblyInformationalVersion("0.6.37-compiled")]
+[assembly: AssemblyVersion("0.7.38")]
+[assembly: AssemblyFileVersion("0.7.38")]
+[assembly: AssemblyInformationalVersion("0.7.38-compiled")]
